chore: move getRelativeWorkspaceUrl out of interal
